Discriminating unions fase 1

+++ ZenthicAlphaAPI/Infrastructure/Roles/Delete/DeleteRoleCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using Application._Common.Events;
-using Application._Common.Exceptions;
+using Application._Common.Failures;
 using Application._Common.Persistence.Databases;
 using Application.Roles.Delete;
+using OneOf;
+using OneOf.Types;
 
 namespace Infrastructure.Roles.Delete;
 
@@ -11,18 +13,22 @@ internal class DeleteRoleCommandHandler(
 )
     : IDeleteRoleCommandHandler
 {
-    public async Task Handle(DeleteRoleCommand command, CancellationToken cancellationToken)
+    public async Task<OneOf<None, Failure>> Handle(DeleteRoleCommand command, CancellationToken cancellationToken)
     {
         var foundRole = await dbContext
             .Roles
-            .FindAsync([command.Id], cancellationToken)
-        ?? throw new NotFoundException(nameof(dbContext.Roles), command.Id);
+            .FindAsync([command.Id], cancellationToken);
+
+        if (foundRole is null)
+            return FailureFactory.NotFound("Role not found", $"No role was found with an Id of {command.Id}");
 
         dbContext.Roles.Remove(foundRole);
         await dbContext.SaveChangesAsync(cancellationToken);
 
         eventPublisher.EnqueueEvent(
-            new RoleDeletedEvent() { Entity = foundRole }
-        );
+                new RoleDeletedEvent() { Entity = foundRole }
+            );
+
+        return new None();
     }
 }
